further updates to database manager source and tests
