+++ spec/server/userControllerIntegrationSpec.js
@@ -1,0 +1,6 @@
+const databaseManagerPath = '../../dataManagement/databaseManager';
+const databaseManagerTestUrl = require("../../../../../private/socialMediaDatabasePrivateTestURL");
+const specDefinition = require('./userControllerBaseSpec');
+
+specDefinition(databaseManagerPath, databaseManagerTestUrl);
+
