+++ spec/server/userControllerUnitSpec.js
@@ -1,0 +1,4 @@
+const databaseManagerPath = '../mock/databaseManager';
+const specDefinition = require('./userControllerBaseSpec');
+
+specDefinition(databaseManagerPath)
